commit date 16/08/2022

+++ example-app/app/Http/Controllers/StudentController.php
@@ -70,15 +70,38 @@ class StudentController extends Controller
             $student->phone = $request->phone;
             $student->qualification_id = implode(',',$request->course);
 
+            // update profile image
             if ($request->hasFile('profile_image')) {
                 $file_type = $request->file('profile_image')->extension();
                 $filename = time().".".$file_type;
-                $image_path = public_path('uploads/students/');
-                unlink($image_path.$student->profile_image);
-                $request->file('profile_image')->move(public_path('uploads/students/'), time() . '.' . $file_type);
-                $student->profile_image = $filename;
+                $image_path = public_path("uploads/students/");
+                if($image_path.$student->profile_image != 0){
+                    unlink($image_path.$student->profile_image);
+                }
+                else{
+                    $request->file('profile_image')->move(public_path('uploads/students/'), time() . '.' . $file_type);
+                    $student->profile_image = $filename;
+                }
             }
-            
+            // multiple image update
+            if($request->hasfile('images'))
+            {
+              foreach (explode(',',($student->images)) as  $value) {
+                  unlink(public_path('uploads/gallery/').$value);
+              }
+  
+              foreach ($request->file('images') as  $value) {
+                $file_type =$value->extension();
+                $filename = uniqid().".".$file_type;
+                $value->move(public_path('uploads/gallery/'),$filename);
+                $galley_image[] = $filename;
+              }
+  
+            //   echo implode(',',$galley_image);
+              $student->images =  implode(',',$galley_image);
+              
+            }
+
             $student->update();
             return redirect('/')->with('message', 'Successfully updated student details');
         }
@@ -87,9 +110,8 @@ class StudentController extends Controller
             // insert new student details to the database
             $request->validate(
                 [
-                    "firstName"=>"required",
-                    "lastName"=> "required",
-                    "email" =>"email | required | unique:students",
+                    "name"=>"required",
+                    "email" =>"email|required | unique:students",
                     "phone" => "required | unique:students",
                     "password" => "required",
                     "profile_image"=>'required | mimes:jpeg,png,jpg | max:2048'
@@ -108,7 +130,18 @@ class StudentController extends Controller
                 $file_path = time().".".$file_type;
                 $request->file('profile_image')->move(public_path('uploads/students'), time() . '.' . $file_type);
                 $student -> profile_image = $file_path;
+            }
+            if($request->hasfile('images'))
+            {
+              foreach ($request->file('images') as  $value) {
+                  $file_type =$value->extension();
+                  $filename = uniqid().".".$file_type;
+                  $value->move(public_path('uploads/gallery/'),$filename);
+                  $galley_image[] = $filename;
               }
+              $student->images =  implode(',',$galley_image);
+              
+            }
               
             $student->save();
             return back()->with('message', 'Student Details Submitted!');
@@ -123,9 +156,10 @@ class StudentController extends Controller
      * @param  \App\Models\student  $student
      * @return \Illuminate\Http\Response
      */
-    public function show(student $student)
+    public function show(student $student, $id)
     {
-
+        $student = Student::find($id);
+        return view('show_details',['student'=>$student]);
     }
 
     /**
@@ -170,6 +204,8 @@ class StudentController extends Controller
         $student->delete();
         $destination = public_path('uploads/students/'. $student->profile_image);
         File::delete($destination);
+        $destination2 = public_path('uploads/uploads/gallery/'.$student->images);
+        File::delete($destination2);
         return back()->with('message', 'Data deleted successfully');
     }
